+++ UserEcho for IOS test app/UserEcho/UserEcho.m
@@ -27,7 +27,7 @@
 }
 
 +(NSString *)version {
-    return @"1.0.4";
+    return @"1.0.5";
 }
 
 @end